WIP untested batched rep pen
