+++ fish_speech_core/lib/models/lm/mod.rs
@@ -1,5 +1,6 @@
 pub mod dual_ar;
 pub mod generate;
+pub mod sampling;
 pub mod utils;
 
 pub use dual_ar::DualARTransformer;
