+++ fish_speech_core/lib/models/lm/utils/mod.rs
@@ -1,3 +1,2 @@
 pub mod encode;
-pub mod sample;
 pub mod text;
