+++ fish_speech_core/src/bin/llama_generate.rs
@@ -4,10 +4,8 @@ use candle_nn::VarBuilder;
 use clap::Parser;
 use fish_speech_core::config::{WhichFishVersion, WhichLM, WhichModel};
 use fish_speech_core::models::lm::generate::generate_blocking;
-use fish_speech_core::models::lm::utils::{
-    encode::PromptEncoder,
-    sample::{load_prompt_text, SamplingArgs},
-};
+use fish_speech_core::models::lm::sampling::SamplingArgs;
+use fish_speech_core::models::lm::utils::encode::{load_prompt_text, PromptEncoder};
 use fish_speech_core::models::lm::{
     dual_ar::{BaseModelArgs, TokenConfig},
     DualARTransformer,
